requirements.txt
matplotlib
numpy
pybullet_planning==0.6.1
pytest
scipy
